+++ backend/src/server/main.py
@@ -96,7 +96,6 @@ class FileStatus(str, enum.Enum):
     COMPLETED = "completed"
     FAILED = "failed"
 
-from sqlalchemy import Column, String, Text, DateTime, Integer, Float
 
 class FileRecord(Base):
     __tablename__ = "files"
@@ -109,36 +108,9 @@ class FileRecord(Base):
     created_at = Column(DateTime, default=datetime.utcnow)
     updated_at = Column(DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)
 
-    # ── New metadata columns ─────────────────────────────────────────────
-    normalized_title     = Column(String, index=True, nullable=True)
-    document_type        = Column(String, index=True, nullable=True)          # invoice, contract, report, ...
-    detected_language    = Column(String, default="en", nullable=True)
-    document_date        = Column(DateTime, nullable=True)
-    year                 = Column(Integer, nullable=True)
-    counterpart          = Column(String, index=True, nullable=True)
-    total_amount         = Column(Float, nullable=True)
-    tags                 = Column(String, nullable=True)                      # comma separated or JSON
-    short_summary        = Column(Text, nullable=True)                        # 200–500 tokens
-
 
 Base.metadata.create_all(bind=engine)
 
-from sqlalchemy import text
-
-def create_document_fts_table():
-    """Create FTS5 virtual table for fast full-text document lookup"""
-    with engine.connect() as conn:
-        conn.execute(text("""
-            CREATE VIRTUAL TABLE IF NOT EXISTS document_fts USING fts5(
-                file_id UNINDEXED,
-                content,
-                tokenize='unicode61'
-            )
-        """))
-        conn.commit()
-
-# Call it once at startup
-create_document_fts_table()
 
 def get_db():
     db = SessionLocal()
@@ -147,36 +119,6 @@ def get_db():
     finally:
         db.close()
 
-
-def index_document_in_fts(
-    file_id: str,
-    filename: str,
-    short_summary: str = "",
-    document_type: str = "",
-    counterpart: str = "",
-    tags: str = "",
-    db: Session
-):
-    """Index document-level searchable content into FTS5"""
-    # Build rich searchable text
-    parts = [
-        filename,
-        unicodedata.normalize("NFKD", filename.lower()).encode("ascii", "ignore").decode(),
-        short_summary.strip(),
-        document_type,
-        counterpart,
-        tags.replace(",", " "),
-    ]
-    searchable_content = " ".join(filter(None, parts))
-
-    db.execute(
-        text("""
-            INSERT OR REPLACE INTO document_fts (file_id, content)
-            VALUES (:file_id, :content)
-        """),
-        {"file_id": file_id, "content": searchable_content}
-    )
-    db.commit()
 
 # ========================= VECTOR STORE =========================
 
@@ -287,221 +229,6 @@ def get_openai_client(model: str = "gemma3") -> AsyncOpenAI:
     return AsyncOpenAI(api_key="http", base_url=DWANI_API_BASE_URL)
 
 
-# utils/token_counter.py
-import tiktoken
-from typing import Optional
-
-# We'll use cl100k_base — most common for recent OpenAI models & many open models
-# If you later switch to a model that uses a different tokenizer, you can make this configurable
-DEFAULT_ENCODING = "cl100k_base"
-
-try:
-    _tokenizer = tiktoken.get_encoding(DEFAULT_ENCODING)
-except Exception as e:
-    raise RuntimeError(f"Failed to load tiktoken encoding '{DEFAULT_ENCODING}': {e}")
-
-def count_tokens(text: str, encoding_name: str = DEFAULT_ENCODING) -> int:
-    """
-    Count the number of tokens in a string using tiktoken.
-    Uses cl100k_base by default (good approximation for most recent models).
-    """
-    if not text:
-        return 0
-    
-    try:
-        return len(_tokenizer.encode(text, disallowed_special=()))
-    except Exception:
-        # Fallback in case of rare encoding errors
-        return len(text) // 4 + len(text.split()) // 2
-
-
-def count_messages_tokens(messages: list[dict], encoding_name: str = DEFAULT_ENCODING) -> int:
-    """
-    Estimate total tokens for a list of chat messages (including role + ~4–8 overhead per message)
-    """
-    total = 0
-    for msg in messages:
-        content = msg.get("content", "")
-        role = msg.get("role", "user")
-        
-        # Rough OpenAI-style overhead
-        overhead = 4 if role == "system" else 3  # system gets slightly more
-        total += count_tokens(content, encoding_name) + overhead
-    
-    # Final reply overhead (~3 tokens)
-    total += 3
-    return total
-
-#====================== RAG EVAL ==============
-
-import json
-from typing import List, Dict, Tuple, Optional
-from openai import AsyncOpenAI
-import logging
-
-logger = logging.getLogger(__name__)
-
-async def llm_judge(
-    prompt: str,
-    client: AsyncOpenAI,
-    model: str = "gemma3",
-    temperature: float = 0.1,
-    max_tokens: int = 350
-) -> Optional[Dict]:
-    """Helper: call LLM and try to parse JSON output"""
-    try:
-        response = await client.chat.completions.create(
-            model=model,
-            messages=[{"role": "user", "content": prompt}],
-            temperature=temperature,
-            max_tokens=max_tokens,
-        )
-        content = response.choices[0].message.content.strip()
-        
-        # Try to clean common markdown / code fence issues
-        if content.startswith("```json"):
-            content = content.split("```json", 1)[1].rsplit("```", 1)[0].strip()
-        elif content.startswith("```"):
-            content = content.split("```", 2)[1].strip()
-            
-        return json.loads(content)
-    except Exception as e:
-        logger.warning(f"LLM judge parsing failed: {e}", exc_info=True)
-        return None
-
-
-async def evaluate_rag_triad(
-    question: str,
-    contexts: List[str],
-    answer: str,
-    client: AsyncOpenAI,
-    model: str = "gemma3"
-) -> Dict[str, Dict[str, float | str]]:
-    """
-    Evaluate RAG response using three reference-free metrics.
-    Returns dict with metric name → {score: float 0–1, explanation: str}
-    """
-    results = {}
-
-    # ── 1. Faithfulness / Groundedness ────────────────────────────────────
-    faithfulness_prompt = f"""You are an impartial evaluator checking if an answer is fully grounded in the given context.
-
-CONTEXT (multiple chunks):
-{chr(10).join(f"[{i+1}] {chunk.strip()}" for i, chunk in enumerate(contexts))}
-
-QUESTION:
-{question}
-
-ANSWER:
-{answer}
-
-Instructions:
-1. Break the ANSWER into individual factual statements/claims.
-2. For each claim, verify if it is directly supported, strongly implied, or entailed by the CONTEXT.
-   - Supported = clearly present (possibly paraphrased)
-   - Partially supported = weak implication
-   - Unsupported / contradicted / invented = hallucination
-3. Score = (number of fully/partially supported claims) / (total claims)
-   - Range: 0.0 (completely hallucinated) to 1.0 (perfectly grounded)
-
-Return **only** valid JSON:
-{{
-  "score": <float between 0.0 and 1.0>,
-  "explanation": "one or two sentence summary of findings",
-  "claim_count": <int>,
-  "supported_count": <int>
-}}
-"""
-
-    faith_result = await llm_judge(faithfulness_prompt, client, model)
-    if faith_result and "score" in faith_result:
-        results["faithfulness"] = {
-            "score": float(faith_result["score"]),
-            "explanation": faith_result.get("explanation", "No explanation provided")
-        }
-    else:
-        results["faithfulness"] = {"score": 0.0, "explanation": "Evaluation failed"}
-
-    # ── 2. Answer Relevancy ───────────────────────────────────────────────
-    relevancy_prompt = f"""Evaluate how relevant and focused the ANSWER is to the QUESTION.
-
-QUESTION:
-{question}
-
-ANSWER:
-{answer}
-
-Scoring guidelines (0.0–1.0):
-• 1.0 = Directly, concisely, and completely answers the question
-• 0.7–0.9 = Mostly relevant, minor digressions
-• 0.4–0.6 = Partially relevant or contains noticeable irrelevant content
-• 0.0–0.3 = Mostly off-topic, evasive, or rambling
-
-Return **only** JSON:
-{{
-  "score": <float 0.0–1.0>,
-  "explanation": "one sentence explanation"
-}}
-"""
-
-    rel_result = await llm_judge(relevancy_prompt, client, model)
-    if rel_result and "score" in rel_result:
-        results["answer_relevancy"] = {
-            "score": float(rel_result["score"]),
-            "explanation": rel_result.get("explanation", "")
-        }
-    else:
-        results["answer_relevancy"] = {"score": 0.0, "explanation": "Evaluation failed"}
-
-    # ── 3. Context Relevancy (average chunk relevance) ─────────────────────
-    if not contexts:
-        results["context_relevancy"] = {"score": 0.0, "explanation": "No context retrieved"}
-    else:
-        chunk_scores = []
-        explanations = []
-
-        for i, chunk in enumerate(contexts, 1):
-            chunk_prompt = f"""Rate how relevant this CONTEXT chunk is to the QUESTION.
-
-QUESTION:
-{question}
-
-CONTEXT CHUNK [{i}/{len(contexts)}]:
-{chunk.strip()}
-
-Score 0.0–1.0:
-• 1.0 = highly relevant, necessary information
-• 0.6–0.9 = useful / related
-• 0.3–0.5 = marginally related
-• 0.0–0.2 = irrelevant / noise
-
-Return **only** JSON:
-{{
-  "score": <float>,
-  "explanation": "short reason"
-}}
-"""
-
-            chunk_eval = await llm_judge(chunk_prompt, client, model, max_tokens=120)
-            if chunk_eval and "score" in chunk_eval:
-                score = float(chunk_eval["score"])
-                chunk_scores.append(score)
-                explanations.append(f"Chunk {i}: {chunk_eval.get('explanation', '—')} ({score:.2f})")
-            else:
-                chunk_scores.append(0.0)
-
-        if chunk_scores:
-            avg_score = sum(chunk_scores) / len(chunk_scores)
-            results["context_relevancy"] = {
-                "score": round(avg_score, 3),
-                "explanation": f"Average over {len(contexts)} chunks • " + " | ".join(explanations[:3])
-            }
-        else:
-            results["context_relevancy"] = {"score": 0.0, "explanation": "Evaluation failed"}
-
-    return results
-
-
 # ========================= SERVICES =========================
 
 async def extract_text_from_images_per_page(images: List[Image.Image]) -> List[str]:
@@ -586,150 +313,6 @@ def generate_pdf_from_text(text: str) -> BytesIO:
 
 # ========================= BACKGROUND TASK =========================
 
-import json
-import logging
-from datetime import datetime
-from typing import Optional
-
-from sqlalchemy.orm import Session
-from openai import AsyncOpenAI
-
-from .models import FileRecord
-from .utils import get_openai_client, clean_text
-from .services import index_document_in_fts   # assuming this function exists
-
-logger = logging.getLogger(__name__)
-
-# ──────────────────────────────────────────────────────────────────────────────
-# LLM METADATA EXTRACTION PROMPT (production-ready, structured output)
-# ──────────────────────────────────────────────────────────────────────────────
-
-METADATA_EXTRACTION_PROMPT = """You are an accurate document metadata extractor. 
-Analyze the provided document text excerpt and return ONLY a valid JSON object.
-
-Rules:
-- Be conservative: only extract information that is clearly and explicitly present.
-- If a field is uncertain or not present → use null (not empty string or "unknown")
-- Do NOT hallucinate dates, amounts, names, or types
-- Infer document_type only if very confident
-- Use ISO 8601 for dates (YYYY-MM-DD) when possible
-- Normalize amounts: remove currency symbols → keep only the numeric value
-- Currency: use ISO 4217 code (USD, EUR, GBP, ...) or null
-- Language: use ISO 639-1 code (en, fr, de, es, it, ...)
-- Summary: 1–3 concise sentences, max ~250 tokens
-
-Return exactly this structure:
-
-{
-  "document_type":      string | null,   // invoice, receipt, contract, bank_statement, report, letter, email, terms_of_service, insurance_policy, tax_form, payslip, ...
-  "title":              string | null,
-  "document_date":      string | null,   // issuance / due / statement / signing date
-  "year":               integer | null,
-  "counterparty":       string | null,   // supplier, client, bank, employer, insurer, ...
-  "total_amount":       number | null,
-  "currency":           string | null,
-  "language":           string,
-  "short_summary":      string
-}
-
-Examples:
-
-Input:
-"INVOICE # INV-2025-0789
-Tesla Energy Inc.
-Date: December 15, 2025
-Due: January 14, 2026
-Powerwall 3 installation - 1 unit
-Total: 12,450.00 EUR"
-
-Output:
-{
-  "document_type": "invoice",
-  "title": "Powerwall 3 Installation Invoice",
-  "document_date": "2025-12-15",
-  "year": 2025,
-  "counterparty": "Tesla Energy Inc.",
-  "total_amount": 12450.00,
-  "currency": "EUR",
-  "language": "en",
-  "short_summary": "Invoice issued by Tesla Energy Inc. on December 15, 2025 for Powerwall 3 installation services totaling 12,450 EUR."
-}
-
-Now extract metadata from this document excerpt:
-
-{excerpt}
-
-Respond ONLY with valid JSON — no explanation, no markdown, no code fences.
-"""
-
-
-async def extract_metadata_with_llm(
-    page_texts: list[str],
-    model: str = "gemma3",           # or "qwen2.5-7b-instruct", "phi-4-mini", etc.
-    max_pages_for_context: int = 4,
-    max_chars: int = 12000
-) -> dict:
-    """
-    Call LLM to extract structured metadata from document text.
-    Returns dict with extracted fields (or empty on failure).
-    """
-    if not page_texts:
-        return {}
-
-    # Take first few pages — usually contains title, date, parties, type
-    excerpt_pages = page_texts[:max_pages_for_context]
-    excerpt = "\n\n".join(excerpt_pages)
-    excerpt = clean_text(excerpt)[:max_chars]
-
-    if len(excerpt.strip()) < 200:
-        logger.warning("Excerpt too short for metadata extraction")
-        return {}
-
-    prompt = METADATA_EXTRACTION_PROMPT.format(excerpt=excerpt)
-
-    client: AsyncOpenAI = get_openai_client(model=model)
-
-    try:
-        response = await client.chat.completions.create(
-            model=model,
-            messages=[{"role": "user", "content": prompt}],
-            temperature=0.1,
-            max_tokens=768,
-            top_p=0.95,
-        )
-
-        content = response.choices[0].message.content.strip()
-
-        # Clean possible markdown / fences
-        if content.startswith("```json"):
-            content = content.split("```json", 1)[1].rsplit("```", 1)[0].strip()
-        elif content.startswith("```"):
-            content = content.split("```", 2)[1].strip()
-
-        meta = json.loads(content)
-
-        # Basic validation / normalization
-        if not isinstance(meta, dict):
-            return {}
-
-        # Ensure expected types
-        meta["year"] = int(meta["year"]) if meta.get("year") else None
-        meta["total_amount"] = float(meta["total_amount"]) if meta.get("total_amount") else None
-
-        return meta
-
-    except json.JSONDecodeError as e:
-        logger.warning(f"Metadata JSON parse failed: {e}", exc_info=True)
-    except Exception as e:
-        logger.error(f"Metadata extraction LLM call failed: {e}", exc_info=True)
-
-    return {}
-
-
-# ──────────────────────────────────────────────────────────────────────────────
-# INTEGRATION INTO BACKGROUND PROCESSING TASK
-# ──────────────────────────────────────────────────────────────────────────────
-
 async def background_extraction_task(file_id: str, pdf_bytes: bytes, filename: str, db: Session):
     file_record = db.query(FileRecord).filter(FileRecord.id == file_id).first()
     if not file_record:
@@ -744,73 +327,20 @@ async def background_extraction_task(file_id: str, pdf_bytes: bytes, filename: s
 
         full_text = "\n\n".join(page_texts)
         file_record.extracted_text = full_text
-
-        # ── Basic fallback values ────────────────────────────────────────
-        normalized = unicodedata.normalize("NFKD", filename.lower()).encode("ascii", "ignore").decode()
-        file_record.normalized_title = normalized.replace(".pdf", "").strip()
-
-        # Default short summary from first page(s)
-        fallback_summary = "\n".join(page_texts[:2])[:800] + "..." if page_texts else ""
-        file_record.short_summary = fallback_summary
-
-        # ── LLM METADATA EXTRACTION ──────────────────────────────────────
-        meta = await extract_metadata_with_llm(page_texts)
-
-        if meta:
-            logger.info(f"LLM metadata extracted for {filename}: {meta.get('document_type')} - {meta.get('title')}")
-
-            file_record.document_type    = meta.get("document_type") or file_record.document_type
-            file_record.normalized_title = meta.get("title") or file_record.normalized_title
-            file_record.short_summary    = meta.get("short_summary") or file_record.short_summary
-            file_record.counterpart      = meta.get("counterparty")
-            file_record.total_amount     = meta.get("total_amount")
-            file_record.currency         = meta.get("currency")
-            file_record.detected_language = meta.get("language", "en")
-
-            # Handle date
-            doc_date_str = meta.get("document_date")
-            if doc_date_str and isinstance(doc_date_str, str) and len(doc_date_str) >= 4:
-                try:
-                    dt = datetime.fromisoformat(doc_date_str)
-                    file_record.document_date = dt
-                    file_record.year = dt.year
-                except ValueError:
-                    pass
-
-            # Use LLM year if date parsing failed but year is present
-            if file_record.year is None and meta.get("year"):
-                try:
-                    file_record.year = int(meta["year"])
-                except:
-                    pass
-
-        # Save updated record
         file_record.status = FileStatus.COMPLETED
         db.commit()
 
-        # ── Store vector embeddings ──────────────────────────────────────
         await store_embeddings_with_pages(file_id, filename, page_texts)
-
-        # ── Index enriched metadata into FTS5 for fast search ────────────
-        index_document_in_fts(
-            file_id=file_id,
-            filename=filename,
-            short_summary=file_record.short_summary or "",
-            document_type=file_record.document_type or "",
-            counterpart=file_record.counterpart or "",
-            tags="",  # extend later if you add tags field
-            db=db
-        )
-
-        logger.info(f"Processing completed for {filename} ({len(page_texts)} pages)")
+        logger.info(f"Extraction completed for {filename} ({len(page_texts)} pages)")
 
     except Exception as e:
         file_record.status = FileStatus.FAILED
         file_record.error_message = str(e)
-        logger.error(f"Extraction failed for {file_id}: {e}", exc_info=True)
+        logger.error(f"Extraction failed for {file_id}: {e}")
     finally:
         file_record.updated_at = datetime.utcnow()
         db.commit()
+
 
 # ========================= HYBRID SEARCH HELPERS =========================
 
@@ -823,48 +353,6 @@ def reciprocal_rank_fusion(results: List[Tuple[int, float]], k: int = 60) -> Lis
     return sorted(score_dict.keys(), key=lambda i: score_dict[i], reverse=True)
 
 
-
-from typing import List, Optional
-
-async def select_relevant_file_ids(
-    question: str,
-    user_selected_file_ids: Optional[List[str]] = None,
-    db: Session,
-    max_files: int = 7,
-) -> List[str]:
-    """
-    Fast first-pass: find most promising documents using FTS5
-    """
-    if user_selected_file_ids and len(user_selected_file_ids) <= max_files:
-        return user_selected_file_ids
-
-    # Prepare FTS5 query
-    words = [w for w in question.lower().split() if len(w) >= 3]
-    if not words:
-        fts_query = question.lower()[:80]
-    else:
-        fts_query = " OR ".join(f'"{w}"' for w in words)
-
-    sql = text("""
-        SELECT file_id
-        FROM document_fts
-        WHERE content MATCH :query
-        ORDER BY rank
-        LIMIT :limit
-    """)
-
-    result = db.execute(sql, {"query": fts_query, "limit": max_files * 2})
-    candidates = [row[0] for row in result.fetchall()]
-
-    # Intersect with user-selected files if provided
-    if user_selected_file_ids:
-        candidates = [fid for fid in candidates if fid in set(user_selected_file_ids)]
-
-    # Fallback: if nothing matched, return all user-selected (or empty)
-    if not candidates and user_selected_file_ids:
-        return user_selected_file_ids[:max_files]
-
-    return candidates[:max_files]
 # ========================= MULTI-DOCUMENT CONTRADICTION DETECTION =========================
 
 async def detect_contradictions(question: str, sources: List[Dict]) -> Optional[str]:
@@ -1027,305 +515,150 @@ def list_files(limit: int = 20, db: Session = Depends(get_db)):
     ]
 
 
-from typing import List, Dict, Optional
-import logging
-from fastapi import HTTPException, Depends, Body
-from sqlalchemy.orm import Session
-from openai import AsyncOpenAI
-from rank_bm25 import BM25Okapi
-from sqlalchemy import text
-
-from .models import FileRecord, FileStatus, MultiChatRequest, ChatMessage
-from .database import get_db, engine
-from .vector_store import collection, embedding_function
-from .utils import count_tokens, count_messages_tokens, reciprocal_rank_fusion
-from .services import detect_contradictions, evaluate_rag_triad, get_openai_client
-
-logger = logging.getLogger(__name__)
-
-# ──────────────────────────────────────────────────────────────────────────────
-# PASS 1: Fast candidate document selection using FTS5
-# ──────────────────────────────────────────────────────────────────────────────
-
-async def select_relevant_file_ids(
-    question: str,
-    user_selected_file_ids: Optional[List[str]] = None,
-    db: Session,
-    max_files: int = 7,
-) -> List[str]:
-    """
-    Fast first-pass: find most promising documents using FTS5 full-text search
-    Returns ranked list of file_ids most likely relevant to the question
-    """
-    if user_selected_file_ids and len(user_selected_file_ids) <= max_files:
-        return user_selected_file_ids
-
-    # Prepare FTS5 query
-    words = [w.strip() for w in question.lower().split() if len(w.strip()) >= 3]
-    if not words:
-        fts_query = question.lower()[:80].strip()
-    else:
-        fts_query = " OR ".join(f'"{w}"' for w in words if w)
-
-    if not fts_query:
-        fts_query = question.lower()[:60]
-
-    sql = text("""
-        SELECT file_id
-        FROM document_fts
-        WHERE content MATCH :query
-        ORDER BY rank
-        LIMIT :limit
-    """)
-
-    try:
-        result = db.execute(sql, {"query": fts_query, "limit": max_files * 2})
-        candidates = [row[0] for row in result.fetchall()]
-    except Exception as e:
-        logger.warning(f"FTS5 query failed: {e}", exc_info=True)
-        candidates = []
-
-    # Intersect with user-provided file_ids if any
-    if user_selected_file_ids:
-        candidates = [fid for fid in candidates if fid in set(user_selected_file_ids)]
-
-    # Fallback: if no matches from FTS5, fall back to all provided files
-    if not candidates and user_selected_file_ids:
-        logger.info("No FTS5 matches → falling back to user-selected files")
-        return user_selected_file_ids[:max_files]
-
-    selected = candidates[:max_files]
-    logger.debug(f"FTS5 selected {len(selected)} candidates: {selected}")
-    return selected
-
-
-# ──────────────────────────────────────────────────────────────────────────────
-# MAIN CHAT ENDPOINT — Two-pass hybrid RAG
-# ──────────────────────────────────────────────────────────────────────────────
-
 @app.post("/chat-with-document", tags=["Files"])
-async def chat_with_documents(
-    request: MultiChatRequest,
-    db: Session = Depends(get_db)
-):
-    """
-    Multi-document RAG chat endpoint with:
-    • Two-pass retrieval (FTS5 candidate selection → filtered hybrid RAG)
-    • Page-aware citations
-    • Basic contradiction detection
-    • Lightweight RAG triad evaluation
-    """
+async def chat_with_documents(request: MultiChatRequest, db: Session = Depends(get_db)):
     if not request.file_ids:
-        raise HTTPException(status_code=400, detail="At least one file_id is required")
+        raise HTTPException(status_code=400, detail="file_ids required")
 
-    # Load and validate file records
     records = db.query(FileRecord).filter(FileRecord.id.in_(request.file_ids)).all()
     if len(records) != len(request.file_ids):
-        raise HTTPException(status_code=404, detail="One or more files not found")
+        raise HTTPException(status_code=404, detail="Some files not found")
 
-    for record in records:
-        if record.status != FileStatus.COMPLETED or not record.extracted_text:
-            raise HTTPException(
-                status_code=400,
-                detail=f"File '{record.filename}' is not yet fully processed"
-            )
+    for r in records:
+        if r.status != FileStatus.COMPLETED:
+            raise HTTPException(status_code=400, detail=f"File {r.filename} not processed")
 
-    # Get latest user question
-    user_messages = [m for m in request.messages if m.role.lower() == "user"]
-    if not user_messages:
-        raise HTTPException(status_code=400, detail="No user question found")
-    question = user_messages[-1].content.strip()
+    user_msg = [m for m in request.messages if m.role == "user"]
+    if not user_msg:
+        raise HTTPException(status_code=400, detail="No user question")
+    question = user_msg[-1].content.lower()
 
-    # ── PASS 1: Candidate document selection ────────────────────────────────
-    candidate_file_ids = await select_relevant_file_ids(
-        question=question,
-        user_selected_file_ids=request.file_ids,
-        db=db,
-        max_files=6           # ← tune this (4–10 is typical sweet spot)
-    )
-
-    if not candidate_file_ids:
-        return {
-            "answer": "No relevant documents matched this question in the provided files.",
-            "sources": [],
-            "usage": {
-                "context_tokens_used": 0,
-                "total_input_tokens_estimated": 0,
-                "max_context_allowed": MAX_CONTEXT_TOKENS
-            },
-            "rag_evaluation": {}
-        }
-
-    logger.info(f"Pass 1 selected {len(candidate_file_ids)} candidate files")
-
-    # ── PASS 2: Hybrid retrieval only on selected candidates ────────────────
+    # === Hybrid Search ===
     vector_results = collection.query(
         query_embeddings=embedding_function([question]),
-        n_results=40,                      # more generous now that files are filtered
-        where={"file_id": {"$in": candidate_file_ids}},
-        include=["documents", "metadatas", "distances"]
+        n_results=20,
+        where={"file_id": {"$in": request.file_ids}},
+        include=["documents", "metadatas", "distances"],
     )
 
-    vector_docs = vector_results["documents"][0] or []
-    vector_metas = vector_results["metadatas"][0] or []
-    vector_distances = vector_results["distances"][0] or []
+    vector_docs = vector_results["documents"][0] if vector_results["documents"] else []
+    vector_metas = vector_results["metadatas"][0] if vector_results["metadatas"] else []
+    vector_distances = vector_results["distances"][0] if vector_results["distances"] else []
 
-    # BM25 re-ranking on retrieved chunks
-    bm25_ranked = []
-    if vector_docs:
-        tokenized_docs = [doc.lower().split() for doc in vector_docs]
-        bm25 = BM25Okapi(tokenized_docs)
-        tokenized_query = question.lower().split()
-        bm25_scores = bm25.get_scores(tokenized_query)
-        bm25_ranked = [
-            (i, score) for i, score in enumerate(bm25_scores) if score > 0
-        ]
-        bm25_ranked.sort(key=lambda x: x[1], reverse=True)
-
-    # Reciprocal Rank Fusion
     vector_ranked = [(i, dist) for i, dist in enumerate(vector_distances)]
-    combined_pairs = vector_ranked + [(i, -score) for i, score in bm25_ranked]
-    fused_indices = reciprocal_rank_fusion(combined_pairs, k=60)[:20]
 
-    # ── Build context (token-aware) ─────────────────────────────────────────
-    context_parts: List[str] = []
-    sources: List[Dict] = []
-    context_tokens_used = 0
-    RESERVED_TOKENS = 1200
-    max_context_tokens = MAX_CONTEXT_TOKENS - RESERVED_TOKENS
+    all_chunks = [d for d in vector_docs]
+    bm25 = BM25Okapi([doc.lower().split() for doc in all_chunks])
+    tokenized_query = question.split()
+    bm25_scores = bm25.get_scores(tokenized_query)
+    bm25_ranked = [(i, -score) for i, score in enumerate(bm25_scores) if score > 0]
+    bm25_ranked.sort(key=lambda x: x[1], reverse=True)
+
+    all_ranked_pairs = vector_ranked + bm25_ranked
+    fused_indices = reciprocal_rank_fusion(all_ranked_pairs)[:20]
+
+    # === Build context with token limit ===
+    context_parts = []
+    sources = []
+    current_token_count = 0
 
     for idx in fused_indices:
         if idx >= len(vector_docs):
             continue
 
-        chunk = vector_docs[idx]
-        chunk_tokens = count_tokens(chunk)
+        doc = vector_docs[idx]
+        estimated_tokens = len(doc.split()) * 1.3
 
-        if context_tokens_used + chunk_tokens > max_context_tokens:
+        if current_token_count + estimated_tokens > MAX_CONTEXT_TOKENS:
             break
 
-        context_parts.append(chunk)
-        context_tokens_used += chunk_tokens
+        context_parts.append(doc)
+        current_token_count += estimated_tokens
 
         meta = vector_metas[idx]
-        filename = meta.get("filename", "Unknown")
+        filename = meta.get("filename", "Unknown Document")
         page_start = meta.get("page_start")
         page_end = meta.get("page_end")
-        page_str = f"Page {page_start}" if page_start == page_end else f"Pages {page_start}–{page_end}"
+        page_citation = f"Page {page_start}"
+        if page_end != page_start:
+            page_citation = f"Pages {page_start}–{page_end}"
 
         sources.append({
             "filename": filename,
-            "page": page_str,
-            "excerpt": chunk.strip()[:600] + ("..." if len(chunk) > 600 else ""),
-            "relevance_score": round(1 - vector_distances[idx], 3)
-                if idx < len(vector_distances) else 0.45
+            "page": page_citation,
+            "excerpt": doc.strip(),
+            "relevance_score": round(1 - vector_distances[idx] if idx < len(vector_distances) else 0.5, 3)
         })
 
-    context_block = "\n\n".join(context_parts) if context_parts else "[No relevant content found]"
+    context = "\n\n".join(context_parts) if context_parts else "No relevant content was found in the document."
 
-    # ── System prompt ───────────────────────────────────────────────────────
-    unique_docs = ", ".join(sorted(set(s["filename"] for s in sources))) or "the documents"
-    system_prompt = f"""You are an expert analyst working with content extracted from: {unique_docs}.
+    # === Multi-document aware system prompt ===
+    doc_list = ", ".join(set(s["filename"] for s in sources))
+    system_prompt = f"""You are an expert assistant analyzing multiple documents: {doc_list}.
 
-Rules:
-• Clearly state which document(s) and page(s) each fact comes from
-• Note any apparent contradictions between documents
-• Use page numbers when available
-• Be concise, accurate and professional
-• If the question cannot be answered from the provided context, reply only: "I don't have sufficient information in the provided documents to answer this question."
+When answering:
+- Always explicitly mention which document each fact comes from (e.g., "According to Contract.pdf...", "The NDA states...").
+- If information differs between documents, clearly note it.
+- Be precise and cite page numbers when available.
+
+If the answer cannot be found, say "I don't know".
 
 Context:
-{context_block}
-"""
+{context}
 
-    # ── Prepare conversation history (token limited) ────────────────────────
+Answer clearly and professionally."""
+
+    # === Trim conversation history ===
     recent_messages = []
-    history_tokens = 0
+    history_token_estimate = 0
 
-    for msg in reversed(request.messages):
-        content_for_count = f"{msg.role}: {msg.content}"
-        msg_tokens = count_tokens(content_for_count) + 10
-        if history_tokens + msg_tokens > MAX_HISTORY_TOKENS:
+    for msg in reversed(request.messages[-12:]):
+        msg_tokens = len(msg.content.split()) * 1.3
+        if history_token_estimate + msg_tokens > MAX_HISTORY_TOKENS:
             break
         recent_messages.insert(0, msg)
-        history_tokens += msg_tokens
+        history_token_estimate += msg_tokens
 
     full_messages = [
         {"role": "system", "content": system_prompt},
         *[{"role": m.role, "content": m.content} for m in recent_messages]
     ]
 
-    total_input_tokens = count_messages_tokens(full_messages)
-
-    # ── Generate answer ─────────────────────────────────────────────────────
-    client = get_openai_client()
     try:
+        client = get_openai_client()
         response = await client.chat.completions.create(
             model="gemma3",
             messages=full_messages,
-            temperature=0.55,
-            max_tokens=1400,
-            top_p=0.92,
+            temperature=0.5,
+            max_tokens=1024,
         )
+
         answer = response.choices[0].message.content.strip()
-    except Exception as e:
-        logger.error("LLM generation failed", extra={"error": str(e), "question": question[:180]})
-        raise HTTPException(status_code=500, detail="Failed to generate answer")
 
-    # ── Optional contradiction detection ────────────────────────────────────
-    contradiction_note = None
-    if len(sources) >= 2:
-        contradiction_note = await detect_contradictions(question, sources[:10])
+        # === Detect contradictions ===
+        contradiction_warning = await detect_contradictions(question, sources[:10])
 
-    final_answer = answer
-    if contradiction_note and "no contradictions" not in contradiction_note.lower():
-        final_answer = f"⚠️ **Possible Contradiction Detected**\n\n{contradiction_note}\n\n**Answer:**\n{answer}"
+        sources.sort(key=lambda x: x["relevance_score"], reverse=True)
 
-    # ── RAG evaluation (optional, lightweight) ──────────────────────────────
-    eval_metrics = {
-        "faithfulness": {"score": None, "explanation": "not evaluated"},
-        "answer_relevancy": {"score": None, "explanation": "not evaluated"},
-        "context_relevancy": {"score": None, "explanation": "not evaluated"}
-    }
+        final_answer = answer
+        if contradiction_warning:
+            final_answer = f"⚠️ **Potential Contradiction Detected**\n\n{contradiction_warning}\n\n**Answer:**\n{answer}"
 
-    try:
-        eval_metrics = await evaluate_rag_triad(
-            question=question,
-            contexts=context_parts,
-            answer=answer,
-            client=client,
-            model="gemma3"
-        )
-    except Exception as eval_err:
-        logger.warning("RAG evaluation failed", exc_info=True)
-
-    # ── Final response ──────────────────────────────────────────────────────
-    sources_sorted = sorted(sources, key=lambda x: x["relevance_score"], reverse=True)
-
-    return {
-        "answer": final_answer,
-        "sources": [
-            {
-                "filename": s["filename"],
-                "page": s["page"],
-                "excerpt": s["excerpt"],
-                "relevance_score": s["relevance_score"]
-            }
-            for s in sources_sorted[:6]   # top 6 most relevant sources
-        ],
-        "usage": {
-            "context_tokens_used": context_tokens_used,
-            "total_input_tokens_estimated": total_input_tokens,
-            "max_context_allowed": MAX_CONTEXT_TOKENS,
-            "selected_documents": len(candidate_file_ids)
-        },
-        "rag_evaluation": eval_metrics,
-        "debug_info": {  # optional — remove in production if desired
-            "candidate_file_ids": candidate_file_ids,
-            "retrieved_chunks": len(vector_docs),
-            "fused_chunks_used": len(context_parts)
+        return {
+            "answer": final_answer,
+            "sources": [
+                {
+                    "filename": s["filename"],
+                    "page": s["page"],
+                    "excerpt": s["excerpt"][:500],
+                    "relevance_score": s["relevance_score"]
+                }
+                for s in sources[:5]
+            ]
         }
-    }
+    except Exception as e:
+        logger.error(f"LLM call failed: {e}")
+        raise HTTPException(status_code=500, detail="Failed to generate response")
 
 
 @app.delete("/files/{file_id}", tags=["Files"])
